Hotfix entity data load eval (#94)
* Hotfix for entity loading

* Hotfix for entity loading - fixed tests

* Hotfix for remving slice dataset loading for dumping

* Added log statements eval

* Added log statements eval

* Changed eval to be dump once then merge files rather than interleaved

* Added size estimate

* Added gc and memory print statments

* Added logs for psutil mem

* Fix setup

* Pass 87 on redo eval for large scale. This time changed how writting to big memmap file with GPUs

* MOre agressive logging

* Maybe it's the memmapping...let's try again

* Sharing file change

* Sharing file unchange

* Allow for overwrite eval data

* Allow for data splitting

* DEbugging for parallel dump

* DEbugging for parallel dump

* Redid subset into dataset for eval

* Remove extra print statements

* Updated changelog

+++ bootleg/dataset.py
@@ -9,6 +9,7 @@ import tempfile
 import time
 import traceback
 import warnings
+from collections import defaultdict
 
 import numpy as np
 import torch
@@ -1248,11 +1249,13 @@ class BootlegDataset(EmmentalDataset):
         name: internal dataset name
         dataset: dataset file
         use_weak_label: whether to use weakly labeled mentions or not
+        load_entity_data: whether to load entity data or not
         tokenizer: sentence tokenizer
         entity_symbols: entity database class
         dataset_threads: number of threads to use
         split: data split
         is_bert: is the tokenizer a BERT or not
+        dataset_range: offset into dataset
     """
 
     def __init__(
@@ -1261,11 +1264,13 @@ class BootlegDataset(EmmentalDataset):
         name,
         dataset,
         use_weak_label,
+        load_entity_data,
         tokenizer,
         entity_symbols,
         dataset_threads,
         split="train",
         is_bert=True,
+        dataset_range=None,
     ):
         """Bootleg dataset initlializer."""
         log_rank_0_info(
@@ -1481,42 +1486,50 @@ class BootlegDataset(EmmentalDataset):
             f"_desc{int(data_config.use_entity_desc)}.bin",
         )
         log_rank_0_debug(logger, f"Seeing if {self.save_entity_dataset_name} exists")
-        if data_config.overwrite_preprocessed_data or (
-            not os.path.exists(self.save_entity_dataset_name)
-        ):
-            st_time = time.time()
-            log_rank_0_info(logger, "Building entity data from scatch.")
-            try:
-                # Creating/saving data
-                build_and_save_entity_inputs(
-                    self.save_entity_dataset_name,
-                    self.X_entity_storage,
-                    data_config,
-                    dataset_threads,
-                    tokenizer,
-                    entity_symbols,
-                )
-                log_rank_0_debug(
-                    logger, f"Finished prepping data in {time.time() - st_time}"
-                )
-            except Exception as e:
-                tb = traceback.TracebackException.from_exception(e)
-                logger.error(e)
-                logger.error(traceback.format_exc())
-                logger.error("\n".join(tb.stack.format()))
-                os.remove(self.save_entity_dataset_name)
-                raise
+        if load_entity_data:
+            if data_config.overwrite_preprocessed_data or (
+                not os.path.exists(self.save_entity_dataset_name)
+            ):
+                st_time = time.time()
+                log_rank_0_info(logger, "Building entity data from scatch.")
+                try:
+                    # Creating/saving data
+                    build_and_save_entity_inputs(
+                        self.save_entity_dataset_name,
+                        self.X_entity_storage,
+                        data_config,
+                        dataset_threads,
+                        tokenizer,
+                        entity_symbols,
+                    )
+                    log_rank_0_debug(
+                        logger, f"Finished prepping data in {time.time() - st_time}"
+                    )
+                except Exception as e:
+                    tb = traceback.TracebackException.from_exception(e)
+                    logger.error(e)
+                    logger.error(traceback.format_exc())
+                    logger.error("\n".join(tb.stack.format()))
+                    os.remove(self.save_entity_dataset_name)
+                    raise
 
-        X_entity_dict = self.build_data_entity_dicts(
-            self.save_entity_dataset_name, self.X_entity_storage
-        )
-        self.X_entity_dict = X_entity_dict
+            X_entity_dict = self.build_data_entity_dicts(
+                self.save_entity_dataset_name, self.X_entity_storage
+            )
+            self.X_entity_dict = X_entity_dict
+        else:
+            self.X_entity_dict = None
 
         log_rank_0_debug(logger, "Removing temporary output files")
         shutil.rmtree(temp_output_folder, ignore_errors=True)
         log_rank_0_info(
             logger,
             f"Final data initialization time for {split} is {time.time() - global_start}s",
+        )
+        self.dataset_range = (
+            list(range(len(X_dict[next(iter(X_dict.keys()))])))
+            if dataset_range is None
+            else dataset_range
         )
         # Set spawn back to original/default, which is "fork" or "spawn".
         # This is needed for the Meta.config to be correctly passed in the collate_fn.
@@ -1608,6 +1621,17 @@ class BootlegDataset(EmmentalDataset):
         X_dict["entity_to_mask"] = torch.from_numpy(mmap_label_file["entity_to_mask"])
         return X_dict
 
+    def get_sentidx_to_rowids(self):
+        """Get mapping from sent idx to row id in X_dict.
+
+        Returns: Dict of sent idx to row id
+        """
+        sentidx2rowids = defaultdict(list)
+        for i, sent_id in enumerate(self.X_dict["sent_idx"]):
+            # Saving/loading dict will convert numeric keys to strings - keep consistent
+            sentidx2rowids[str(sent_id.item())].append(i)
+        return dict(sentidx2rowids)
+
     def __getitem__(self, index):
         r"""Get item by index.
 
@@ -1616,6 +1640,7 @@ class BootlegDataset(EmmentalDataset):
         Returns:
           Tuple[Dict[str, Any], Dict[str, Tensor]]: Tuple of x_dict and y_dict
         """
+        index = self.dataset_range[index]
         x_dict = {name: feature[index] for name, feature in self.X_dict.items()}
         y_dict = {name: label[index] for name, label in self.Y_dict.items()}
 
@@ -1625,30 +1650,31 @@ class BootlegDataset(EmmentalDataset):
             x_dict["input_ids"] = input_ids
         # Get the entity_cand_eid
         entity_cand_eid = self.alias2cands_model(x_dict["alias_idx"]).long()
-        entity_cand_input_ids = []
-        entity_cand_token_type_ids = []
-        entity_cand_attention_mask = []
-        # Get the entity token ids
-        for eid in entity_cand_eid:
-            if self.split == "train" and self.popularity_mask:
-                entity_input_ids = self._mask_entity_input_ids(x_dict, eid)
-            else:
-                entity_input_ids = self.X_entity_dict["entity_input_ids"][eid]
-            entity_cand_input_ids.append(entity_input_ids)
-            entity_cand_token_type_ids.append(
-                self.X_entity_dict["entity_token_type_ids"][eid]
+        if self.X_entity_dict is not None:
+            entity_cand_input_ids = []
+            entity_cand_token_type_ids = []
+            entity_cand_attention_mask = []
+            # Get the entity token ids
+            for eid in entity_cand_eid:
+                if self.split == "train" and self.popularity_mask:
+                    entity_input_ids = self._mask_entity_input_ids(x_dict, eid)
+                else:
+                    entity_input_ids = self.X_entity_dict["entity_input_ids"][eid]
+                entity_cand_input_ids.append(entity_input_ids)
+                entity_cand_token_type_ids.append(
+                    self.X_entity_dict["entity_token_type_ids"][eid]
+                )
+                entity_cand_attention_mask.append(
+                    self.X_entity_dict["entity_attention_mask"][eid]
+                )
+            # Create M x K x token length
+            x_dict["entity_cand_input_ids"] = torch.stack(entity_cand_input_ids, dim=0)
+            x_dict["entity_cand_token_type_ids"] = torch.stack(
+                entity_cand_token_type_ids, dim=0
             )
-            entity_cand_attention_mask.append(
-                self.X_entity_dict["entity_attention_mask"][eid]
+            x_dict["entity_cand_attention_mask"] = torch.stack(
+                entity_cand_attention_mask, dim=0
             )
-        # Create M x K x token length
-        x_dict["entity_cand_input_ids"] = torch.stack(entity_cand_input_ids, dim=0)
-        x_dict["entity_cand_token_type_ids"] = torch.stack(
-            entity_cand_token_type_ids, dim=0
-        )
-        x_dict["entity_cand_attention_mask"] = torch.stack(
-            entity_cand_attention_mask, dim=0
-        )
         x_dict["entity_cand_eval_mask"] = entity_cand_eid == -1
         # Handles the index errors with -1 indexing into an embedding
         x_dict["entity_cand_eid"] = torch.where(
@@ -1770,6 +1796,10 @@ class BootlegDataset(EmmentalDataset):
             f"Bootleg Dataset. Data at {self.save_dataset_name}. "
             f"Labels at {self.save_labels_name}. "
         )
+
+    def __len__(self):
+        """Length."""
+        return len(self.dataset_range)
 
 
 class BootlegEntityDataset(EmmentalDataset):
